fixes issue when output dir does not exists and HOT is invoked
As per @blanklob tests this issue was occurring when attempting to develop when an empty output directory existed. This defect was using older logic and globs were not being resolved.

This commit fixes that and when empty output directories exist and user invokes HOT mode then all related "alive" layouts will publish and inject accordingly, regardless of whether or not an empty output exists.